Slice and batch nhwc to nchw added
